Organise file system
Relates #21
